Pulled Pokemon Games From API Into Dropdown
Moved API calls to Home component and added API call to get pokemon games and populated dropdown
